# openapi.json
{
  "components": {
    "schemas": {
...
      "AddTrackedArtistsRequest": {
        "properties": {
          "artist_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Artist Ids",
            "type": "array"
          }
        },
        "title": "AddTrackedArtistsRequest",
        "type": "object"
      },
      "AddTrackedArtistsResponse": {
        "properties": {
          "added": {
            "title": "Added",
            "type": "integer"
          },
          "already_tracked": {
            "title": "Already Tracked",
            "type": "integer"
          }
        },
        "required": [
          "added",
          "already_tracked"
        ],
        "title": "AddTrackedArtistsResponse",
        "type": "object"
      },
...
      "BucketResponse": {
        "properties": {
          "children": {
            "items": {
              "$ref": "#/components/schemas/BucketResponse"
            },
            "title": "Children",
            "type": "array"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_done": {
            "title": "Is Done",
            "type": "boolean"
          },
          "is_public": {
            "default": false,
            "title": "Is Public",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BucketItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "kind": {
            "default": "review",
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "research_mode": {
            "default": "off",
            "title": "Research Mode",
            "type": "string"
          },
          "type": {
            "default": "general",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "position",
          "is_done"
        ],
        "title": "BucketResponse",
        "type": "object"
      },
      "BucketsResponse": {
        "properties": {
          "buckets": {
            "items": {
              "$ref": "#/components/schemas/BucketResponse"
            },
            "title": "Buckets",
            "type": "array"
          }
        },
        "title": "BucketsResponse",
        "type": "object"
      },
...
      "CreateBucketRequest": {
        "properties": {
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "type": {
            "default": "general",
            "enum": [
              "general",
              "artist"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateBucketRequest",
        "type": "object"
      },
      "CreateGenreRequest": {
        "properties": {
          "definition_md": {
            "default": "",
            "title": "Definition Md",
            "type": "string"
          },
          "label": {
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "slug": {
            "minLength": 1,
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "label"
        ],
        "title": "CreateGenreRequest",
        "type": "object"
      },
...
      "DailyPickItem": {
        "properties": {
          "album_id": {
            "title": "Album Id",
            "type": "string"
          },
          "artist": {
            "title": "Artist",
            "type": "string"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "pick_date": {
            "format": "date",
            "title": "Pick Date",
            "type": "string"
          },
          "spotify_track_id": {
            "title": "Spotify Track Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "track_id": {
            "title": "Track Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "pick_date",
          "track_id",
          "album_id",
          "title",
          "artist",
          "spotify_track_id",
          "created_at",
          "updated_at"
        ],
        "title": "DailyPickItem",
        "type": "object"
      },
...
      "GenreNode": {
        "properties": {
          "album_count": {
            "default": 0,
            "title": "Album Count",
            "type": "integer"
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/GenreNode"
            },
            "title": "Children",
            "type": "array"
          },
          "definition_md": {
            "default": "",
            "title": "Definition Md",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "position": {
            "default": 0,
            "title": "Position",
            "type": "integer"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "label"
        ],
        "title": "GenreNode",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IntegrationsResponse": {
        "properties": {
          "integrations": {
            "items": {
              "$ref": "#/components/schemas/IntegrationResponse"
            },
            "title": "Integrations",
            "type": "array"
          }
        },
        "required": [
          "integrations"
        ],
        "title": "IntegrationsResponse",
        "type": "object"
      },
...
      "MeResponse": {
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "handle": {
            "title": "Handle",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "handle",
          "display_name",
          "created_at"
        ],
        "title": "MeResponse",
        "type": "object"
      },
...
      "PostListResponse": {
        "properties": {
          "posts": {
            "items": {
              "$ref": "#/components/schemas/PostListItem"
            },
            "title": "Posts",
            "type": "array"
          }
        },
        "required": [
          "posts"
        ],
        "title": "PostListResponse",
        "type": "object"
      },
...
      "TrackedArtistResponse": {
        "properties": {
          "added_at": {
            "format": "date-time",
            "title": "Added At",
            "type": "string"
          },
          "artist_id": {
            "format": "uuid",
            "title": "Artist Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          }
        },
        "required": [
          "artist_id",
          "name",
          "added_at"
        ],
        "title": "TrackedArtistResponse",
        "type": "object"
      },
...
      "UpdateMeRequest": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "handle": {
            "anyOf": [
              {
                "pattern": "^[a-z0-9][a-z0-9_-]{1,28}[a-z0-9]$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Handle"
          }
        },
        "title": "UpdateMeRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WritePostRequest": {
        "properties": {
          "album_classics": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Album Classics",
            "type": "object"
          },
          "album_ids": {
            "items": {
              "type": "string"
            },
            "title": "Album Ids",
            "type": "array"
          },
          "artist_ids": {
            "items": {
              "type": "string"
            },
            "title": "Artist Ids",
            "type": "array"
          },
          "body_mdx": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Mdx"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "genre_ids": {
            "items": {
              "type": "string"
            },
            "title": "Genre Ids",
            "type": "array"
          },
          "posted_date": {
            "format": "date",
            "title": "Posted Date",
            "type": "string"
          },
          "rating": {
            "anyOf": [
              {
                "maximum": 5.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "recommended_track_ids": {
            "items": {
              "type": "string"
            },
            "title": "Recommended Track Ids",
            "type": "array"
          },
          "search_index": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Search Index"
          },
          "status": {
            "default": "published",
            "enum": [
              "draft",
              "published",
              "archived"
            ],
            "title": "Status",
            "type": "string"
          },
          "subject_best_new": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Best New"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "WritePostRequest",
        "type": "object"
      },
      "WritePostResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug"
        ],
        "title": "WritePostResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Blog Backend (Lambda/Mangum)",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/buckets": {
      "get": {
        "operationId": "list_buckets_api_buckets_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BucketsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Buckets",
        "tags": [
          "buckets"
        ]
      },
      "post": {
        "operationId": "create_bucket_api_buckets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBucketRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BucketResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Bucket",
        "tags": [
          "buckets"
        ]
      }
    },
...
    "/api/genres": {
      "post": {
        "operationId": "create_genre_api_genres_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGenreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenreNode"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Genre",
        "tags": [
          "genres"
        ]
      }
    },
...
    "/api/integrations": {
      "get": {
        "operationId": "list_integrations_api_integrations_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Integrations",
        "tags": [
          "integrations"
        ]
      }
    },
...
    "/api/me": {
      "delete": {
        "operationId": "delete_me_api_me_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Me",
        "tags": [
          "me"
        ]
      },
      "get": {
        "operationId": "get_me_api_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Me",
        "tags": [
          "me"
        ]
      },
      "patch": {
        "operationId": "update_me_api_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Me",
        "tags": [
          "me"
        ]
      }
    },
...
    "/api/me/tracked-artists": {
      "get": {
        "operationId": "list_tracked_artists_api_me_tracked_artists_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TrackedArtistResponse"
                  },
                  "title": "Response List Tracked Artists Api Me Tracked Artists Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Tracked Artists",
        "tags": [
          "tracked-artists"
        ]
      },
      "post": {
        "operationId": "add_tracked_artists_api_me_tracked_artists_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTrackedArtistsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddTrackedArtistsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Tracked Artists",
        "tags": [
          "tracked-artists"
        ]
      }
    },
...
    "/api/posts": {
      "get": {
        "operationId": "list_posts_api_posts_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Posts",
        "tags": [
          "posts"
        ]
      },
      "post": {
        "operationId": "create_post_api_posts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WritePostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WritePostResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Slug derived from title already exists"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Post",
        "tags": [
          "posts"
        ]
      }
    },
...
    "/api/publish": {
      "post": {
        "operationId": "create_post_api_publish_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Post",
        "tags": [
          "publish"
        ]
      }
    },
...
    "/api/todays-pick": {
      "delete": {
        "description": "Clear today's pick (\"unpost today\"). 204 if a pick was removed; 404 if\nnothing was posted today (re-PUT is the normal path to replace a pick).",
        "operationId": "delete_todays_pick_api_todays_pick_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Todays Pick",
        "tags": [
          "todays-pick"
        ]
      },
      "get": {
        "description": "Today's pick, or null on a no-pick day (home tile hides on null).",
        "operationId": "get_todays_pick_api_todays_pick_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/DailyPickItem"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Todays Pick Api Todays Pick Get"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Todays Pick",
        "tags": [
          "todays-pick"
        ]
      },
      "put": {
        "description": "Set today's pick. Upserts on pick_date = today; re-POST overwrites. The\nserver pins pick_date to today \u2014 the body carries the track's ids + display.",
        "operationId": "put_todays_pick_api_todays_pick_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertTodaysPickRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyPickItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Put Todays Pick",
        "tags": [
          "todays-pick"
        ]
      }
    },
...
  }
}